tests: factor out `assert_downloaded` to DownloadTestHelper.
